Removed references to Jabbr.Models due to move of ContentProviderResultModel
