+++ JabbR/ContentProviders/EmbedContentProvider.cs
@@ -3,7 +3,6 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Net;
 using System.Text.RegularExpressions;
-using JabbR.Models;
 
 namespace JabbR.ContentProviders
 {
