+++ JabbR/ContentProviders/ImageContentProvider.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Net;
-using JabbR.Models;
 
 namespace JabbR.ContentProviders
 {
