+++ JabbR/ContentProviders/ImgurContentProvider.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Linq;
 using System.Net;
-using JabbR.Models;
 
 namespace JabbR.ContentProviders
 {
