+++ JabbR/ContentProviders/TweetContentProvider.cs
@@ -3,7 +3,6 @@ using System.Linq;
 using System.Net;
 using System.Text.RegularExpressions;
 using System.Web;
-using JabbR.Models;
 
 namespace JabbR.ContentProviders
 {
